Expose executeShellCommand, RunCommandOptions, and fetchContents from the node package
