+++ plugins/scaffolder-backend/api-report.md
@@ -3,8 +3,6 @@
 > Do not edit this file. It is a report generated by [API Extractor](https://api-extractor.com/).
 
 ```ts
-/// <reference types="node" />
-
 import { ActionContext as ActionContext_2 } from '@backstage/plugin-scaffolder-node';
 import { CatalogApi } from '@backstage/catalog-client';
 import { CatalogProcessor } from '@backstage/plugin-catalog-node';
@@ -13,7 +11,10 @@ import { Config } from '@backstage/config';
 import { createPullRequest } from 'octokit-plugin-create-pull-request';
 import { Duration } from 'luxon';
 import { Entity } from '@backstage/catalog-model';
+import { executeShellCommand as executeShellCommand_2 } from '@backstage/plugin-scaffolder-node';
+import { ExecuteShellCommandOptions } from '@backstage/plugin-scaffolder-node';
 import express from 'express';
+import { fetchContents as fetchContents_2 } from '@backstage/plugin-scaffolder-node';
 import { GithubCredentialsProvider } from '@backstage/integration';
 import { HumanDuration } from '@backstage/types';
 import { IdentityApi } from '@backstage/plugin-auth-node';
@@ -34,7 +35,6 @@ import { RESOURCE_TYPE_SCAFFOLDER_TEMPLATE } from '@backstage/plugin-scaffolder-
 import { Schema } from 'jsonschema';
 import { ScmIntegrationRegistry } from '@backstage/integration';
 import { ScmIntegrations } from '@backstage/integration';
-import { SpawnOptionsWithoutStdio } from 'child_process';
 import { TaskSecrets as TaskSecrets_2 } from '@backstage/plugin-scaffolder-node';
 import { TaskSpec } from '@backstage/plugin-scaffolder-common';
 import { TaskSpecV1beta3 } from '@backstage/plugin-scaffolder-common';
@@ -43,7 +43,6 @@ import { TemplateActionOptions } from '@backstage/plugin-scaffolder-node';
 import { TemplateEntityStepV1beta3 } from '@backstage/plugin-scaffolder-common';
 import { TemplateParametersV1beta3 } from '@backstage/plugin-scaffolder-common';
 import { UrlReader } from '@backstage/backend-common';
-import { Writable } from 'stream';
 import { ZodType } from 'zod';
 import { ZodTypeDef } from 'zod';
 
@@ -816,17 +815,11 @@ export type DatabaseTaskStoreOptions = {
   database: PluginDatabaseManager | Knex;
 };
 
-// @public
-export const executeShellCommand: (options: RunCommandOptions) => Promise<void>;
+// @public @deprecated
+export const executeShellCommand: typeof executeShellCommand_2;
 
-// @public
-export function fetchContents(options: {
-  reader: UrlReader;
-  integrations: ScmIntegrations;
-  baseUrl?: string;
-  fetchUrl?: string;
-  outputPath: string;
-}): Promise<void>;
+// @public @deprecated
+export const fetchContents: typeof fetchContents_2;
 
 // @public (undocumented)
 export type OctokitWithPullRequestPluginClient = Octokit & {
@@ -876,13 +869,8 @@ export interface RouterOptions {
   taskWorkers?: number;
 }
 
-// @public (undocumented)
-export type RunCommandOptions = {
-  command: string;
-  args: string[];
-  options?: SpawnOptionsWithoutStdio;
-  logStream?: Writable;
-};
+// @public @deprecated
+export type RunCommandOptions = ExecuteShellCommandOptions;
 
 // @public (undocumented)
 export class ScaffolderEntitiesProcessor implements CatalogProcessor {
