+++ plugins/scaffolder-backend/src/deprecated.ts
@@ -19,6 +19,9 @@ import {
   createTemplateAction as createTemplateActionNode,
   TaskSecrets as TaskSecretsNode,
   TemplateAction as TemplateActionNode,
+  executeShellCommand as executeShellCommandNode,
+  ExecuteShellCommandOptions as ExecuteShellCommandOptionsNode,
+  fetchContents as fetchContentsNode,
 } from '@backstage/plugin-scaffolder-node';
 import { JsonObject } from '@backstage/types';
 
@@ -47,3 +50,28 @@ export type TaskSecrets = TaskSecretsNode;
  */
 export type TemplateAction<TInput extends JsonObject> =
   TemplateActionNode<TInput>;
+
+/**
+ * Options for {@link executeShellCommand}.
+ *
+ * @public
+ * @deprecated Use `ExecuteShellCommandOptions` from `@backstage/plugin-scaffolder-node` instead
+ */
+export type RunCommandOptions = ExecuteShellCommandOptionsNode;
+
+/**
+ * Run a command in a sub-process, normally a shell command.
+ *
+ * @public
+ * @deprecated Use `executeShellCommand` from `@backstage/plugin-scaffolder-node` instead
+ */
+export const executeShellCommand = executeShellCommandNode;
+
+/**
+ * A helper function that reads the contents of a directory from the given URL.
+ * Can be used in your own actions, and also used behind fetch:template and fetch:plain
+ *
+ * @public
+ * @deprecated Use `fetchContents` from `@backstage/plugin-scaffolder-node` instead
+ */
+export const fetchContents = fetchContentsNode;
