+++ plugins/scaffolder-backend/src/scaffolder/actions/builtin/fetch/plain.test.ts
@@ -21,9 +21,9 @@ import { resolve as resolvePath } from 'path';
 import { getVoidLogger, UrlReader } from '@backstage/backend-common';
 import { ConfigReader } from '@backstage/config';
 import { ScmIntegrations } from '@backstage/integration';
+import { fetchContents } from '@backstage/plugin-scaffolder-node';
 import { createFetchPlainAction } from './plain';
 import { PassThrough } from 'stream';
-import { fetchContents } from './helpers';
 
 describe('fetch:plain', () => {
   const integrations = ScmIntegrations.fromConfig(
