+++ plugins/scaffolder-backend/src/scaffolder/actions/builtin/fetch/plain.ts
@@ -16,8 +16,10 @@
 
 import { UrlReader, resolveSafeChildPath } from '@backstage/backend-common';
 import { ScmIntegrations } from '@backstage/integration';
-import { fetchContents } from './helpers';
-import { createTemplateAction } from '@backstage/plugin-scaffolder-node';
+import {
+  createTemplateAction,
+  fetchContents,
+} from '@backstage/plugin-scaffolder-node';
 
 /**
  * Downloads content and places it in the workspace, or optionally
