+++ plugins/scaffolder-backend/src/scaffolder/actions/builtin/index.ts
@@ -21,6 +21,3 @@ export * from './fetch';
 export * from './filesystem';
 export * from './publish';
 export * from './github';
-
-export { executeShellCommand } from './helpers';
-export type { RunCommandOptions } from './helpers';
