+++ plugins/scaffolder-node/src/actions/index.ts
@@ -18,4 +18,9 @@ export {
   createTemplateAction,
   type TemplateActionOptions,
 } from './createTemplateAction';
+export {
+  executeShellCommand,
+  type ExecuteShellCommandOptions,
+} from './executeShellCommand';
+export { fetchContents, fetchFile } from './fetch';
 export { type ActionContext, type TemplateAction } from './types';
